@@ -1,3 +1,8 @@
-import VueFroala from './vue-froala'
+import {Froala, FroalaView} from './vue-froala'
 
-export default VueFroala
+export default {
+    install: (app, options) => {
+        app.component('Froala', Froala);
+        app.component('FroalaView', FroalaView);
+    }
+}
